04082026 :Clean Architecture Vertical Slice Domain Model CQRS سبک PostgreSQL با EF Core مدل‌های Developer، Repository، Issue، Skill و Recommendation GitHub API Client همگام‌سازی اولیه Repositoryها همگام‌سازی اولیه Issueها موتور امتیازدهی چندمعیاره ذخیره پیشنهادهای روزانه
در BasicMatchingEngine معیارهای مناسبی در نظر گرفته شده‌اند:

تطابق مهارت
آشنایی با Repository
سابقه مشارکت
تازگی فعالیت پروژه
علاقه و ترجیحات
تاریخچه پیشنهاد
تناسب سطح دشواری

+++ DevMatch.Domain/ValueObjects/MatchScore.cs
@@ -18,14 +18,11 @@ namespace DevMatch.Domain.ValueObjects
     //پس خروجی نباید فقط یک عدد باشد.
     public sealed record MatchScore
     {
-        public int Percentage { get; }
+        public double Percentage { get; }
 
-        public MatchScore(int percentage)
+        public MatchScore(double percentage)
         {
-            Percentage = Math.Clamp(
-                percentage,
-                0,
-                100);
+            Percentage = Math.Clamp(percentage, 0, 100);
         }
 
         public bool IsQualified
